refactoring progress
